python-ai-service/requirements.txt
flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0
spacy==3.7.2
scikit-learn==1.3.0
pandas==2.1.0
numpy==1.25.2
tensorflow==2.13.0
transformers==4.33.2
python-dotenv==1.0.0
requests==2.31.0 